+++ frontend/src/App.jsx
@@ -1,23 +1,13 @@
 import './App.css';
-import './assets/css/app.css';
-// import WebHome from "./assets/components/frontend/WebHome";
-// import Home from "./assets/components/admin/Home";
-// import LoginForm from "./assets/components/Auth/LoginForm";
-// import SignupForm from "./assets/components/Auth/SignupForm";
-// import ForgotPassword from "./assets/components/Auth/ForgotPassword";
-// import ResetPassword from './assets/components/Auth/ResetPassword';
-import NotFound from './assets/components/frontend/NotFound';
+// import './assets/css/app.css';
+import WebHome from "./assets/components/frontend/WebHome";
+
+
 
 const App = () => {
   return (
     <>
-      {/* <WebHome /> */}
-      {/* <Home/> */}
-      {/* <LoginForm /> */}
-      {/* <SignupForm /> */}
-      {/* <ForgotPassword/> */}
-      {/* <ResetPassword/> */}
-      <NotFound/>
+      <WebHome />
     </>
   );
 };
