--- frontend/src/assets/components/frontend/Header.jsx
@@ -1,10 +1,0 @@
-
-const Header = () => {
-  return (
-    <header className="text-center py-1 text-white">
-      Sign up and get 20% off to your first order.
-    </header>
-  );
-};
-
-export default Header;
