+++ frontend/src/assets/components/frontend/WebHome.jsx
@@ -1,12 +1,14 @@
-import "../../../App.css";
-import Header from "./Header";
-import Navigationbar from "./Navigationbar";
+// import "../../../App.css";
+import Header from "./header/Header";
+import Navigationbar from "./navbar/Navigationbar";
+import Nav from './navbar2/Navbar2';
 
 const WebHome = () => {
   return (
     <>
       <Header />
       <Navigationbar />
+      <Nav />
     </>
   );
 };
