+++ frontend/src/assets/components/frontend/header/Header.jsx
@@ -1,0 +1,11 @@
+import './header.css';
+
+const Header = () => {
+  return (
+    <header className="header">
+      <p className="text-center py-2">Sign up and get 20% off to your first order.</p>
+    </header>
+  );
+};
+
+export default Header;
